+++ assets/js/panelbar.js
@@ -8,8 +8,38 @@ if (!jQuery) {
 
 } else {
   $(function() {
+
+    // Visibility toggle & flip
     $(".panelbar__switch").on("click", function () {
-      $(".panelbar, .panelbar__switch").toggleClass("hidden");
+      $(".panelbar").toggleClass("hidden");
     });
+
+    $(".panelbar__flip").on("click", function () {
+      $(".panelbar").toggleClass("top bottom");
+    });
+
+    // Element: toggle
+    if (enhancedJS === true) {
+      $(".panelbar--toggle > a").on("click", function (e) {
+        e.preventDefault();
+
+        var status = $(this).find('span').text() == 'Visible' ? 'hide' : 'publish';
+        var url    = siteURL + "/panel/api/pages/" + status + "/" + currentURI;
+
+        $.ajax({
+          type: "POST",
+          url: url
+        });
+
+        $(this).find('.fa').toggleClass('fa-toggle-off fa-toggle-on');
+        $(this).find('span').text(status == "hide" ? "Invisible" : "Visible");
+
+        setTimeout(function() {
+          location.reload();
+        }, 100);
+
+      });
+    }
+
   });
 }