+++ panel-bar.php
@@ -4,6 +4,7 @@ class PanelBar {
 
   public  $defaults  = array(
                         'panel',
+                        'add',
                         'edit',
                         'toggle',
                         'languages',
@@ -24,7 +25,9 @@ class PanelBar {
                         '__construct',
                         '__output',
                         '__content',
+                        '__controlBtn',
                         '__switchBtn',
+                        '__flipBtn',
                         '__float',
                         '__getCSS',
                         '__getJS',
@@ -81,7 +84,7 @@ class PanelBar {
     if ($user = site()->user() and $user->hasPanelAccess()) {
       $bar  = '<div class="panelbar '.$this->position.' '.($hidden === true ? 'hidden' : '').'" id="panelbar">'.$this->__content().'</div>';
 
-      $bar .= $this->__switchBtn($hidden);
+      $bar .= $this->__controlBtn($hidden);
 
       if ($css) $bar .= $this->__getCSS();
       if ($js)  $bar .= $this->__getJS();
@@ -111,12 +114,29 @@ class PanelBar {
     return $content;
   }
 
-  protected function __switchBtn($hidden = false) {
-    $switch  = '<div class="panelbar__switch '.($hidden === true ? 'hidden' : '').'" id ="panelbar_switch">';
+  protected function __controlBtn() {
+    $controls  = '<div class="panelbar__controls">';
+    $controls .= $this->__flipBtn();
+    $controls .= $this->__switchBtn();
+    $controls .= '</div>';
+    return $controls;
+  }
+
+  protected function __switchBtn() {
+    $switch  = '<div class="panelbar__switch" id ="panelbar_switch">';
     $switch .= '<i class="fa fa-times-circle panelbar__switch--visible"></i>';
     $switch .= '<i class="fa fa-plus-circle panelbar__switch--hidden"></i>';
+    $switch .= '<i class="fa fa-circle panelbar__switch--bg"></i>';
     $switch .= '</div>';
     return $switch;
+  }
+
+  protected function __flipBtn() {
+    $flip  = '<div class="panelbar__flip">';
+    $flip .= '<i class="fa fa-arrow-circle-up panelbar__flip--top"></i>';
+    $flip .= '<i class="fa fa-arrow-circle-down panelbar__flip--bottom"></i>';
+    $flip .= '</div>';
+    return $flip;
   }
 
   protected static function __float($args) {
@@ -127,12 +147,14 @@ class PanelBar {
 
   protected function __getCSS($position = null) {
     $style  = tpl::load(__DIR__ . DS . 'assets' . DS . 'css' . DS . 'panelbar.min.css');
-    $style .= 'body {margin-'.(is_null($position) ? $this->position : $position).': 50px !important}';
     return '<style>'.$style.'</style>';
   }
 
   protected function __getJS() {
     $script  = tpl::load(__DIR__ . DS . 'assets' . DS . 'js' . DS . 'panelbar.min.js');
+    $script .= 'siteURL = "'.$this->site->url().'";';
+    $script .= 'currentURI = "'.$this->page->uri().'";';
+    $script .= 'enhancedJS ='.(c::get('panelbar.enhancedJS', false) ? 'true' : 'false').';';
     return '<script>'.$script.'</script>';
   }
 
@@ -140,12 +162,21 @@ class PanelBar {
 
   /* Public helpers to build elements */
 
+  // TODO: remove redundancy from helpers (esp. in regard to responsiveness)
+
   public static function link($args) {
     $class  = 'panelbar__btn '.self::__float($args).' panelbar--'.$args['id'];
     $block  = '<div class="'.$class.'">';
     $block .= '<a href="'.$args['url'].'">';
-    if (isset($args['icon'])) $block .= '<i class="fa fa-'.$args['icon'].'"></i>';
-    if (isset($args['text'])) $block .= '<span>'.$args['text'].'</span>';
+
+    if (isset($args['icon'])) {
+      $block .= '<i class="fa fa-'.$args['icon'].' '.((isset($args['mobile']) and $args['mobile'] == 'label') ? 'not-mobile' : '').'"></i>';
+    }
+
+    if (isset($args['label'])) {
+      $block .= '<span '.((!isset($args['mobile']) or $args['mobile'] == 'icon') ? 'class="not-mobile"' : '').'>'.$args['label'].'</span>';
+    }
+
     $block .= '</a>';
     $block .= '</div>';
     return $block;
@@ -157,15 +188,23 @@ class PanelBar {
 
     // label
     $block .= '<span>';
-    if (isset($args['icon'])) $block .= '<i class="fa fa-'.$args['icon'].'"></i>';
-    $block .= '<span>'.$args['label'].'</span>';
-    $block .= '<i class="fa fa-caret-'.(c::get('panelbar.position') == 'bottom' ? 'up' : 'down').' fa-styleless"></i>';
+
+    if (isset($args['icon'])) {
+      $block .= '<i class="fa fa-'.$args['icon'].' '.((isset($args['mobile']) and $args['mobile'] == 'label') ? 'not-mobile' : '').'"></i>';
+    }
+
+    if (isset($args['label'])) {
+      $block .= '<span '.((!isset($args['mobile']) or $args['mobile'] == 'icon') ? 'class="not-mobile"' : '').'>'.$args['label'].'</span>';
+    }
+
+    $block .= '<i class="fa fa-caret-up fa-styleless"></i>';
+    $block .= '<i class="fa fa-caret-down fa-styleless"></i>';
     $block .= '</span>';
 
     // all items
     $block .= '<div class="panelbar__dropitems">';
     foreach($args['items'] as $item) {
-      $block .= '<a href="'.$item['url'].'" class="panelbar__dropitem">'.$item['text'].'</a>';
+      $block .= '<a href="'.$item['url'].'" class="panelbar__dropitem">'.$item['label'].'</a>';
     }
     $block .= '</div>';
 
@@ -179,48 +218,63 @@ class PanelBar {
 
   protected function panel() {
     return self::link(array(
-      'id'   => 'panel',
-      'icon' => 'cogs',
-      'url'  => site()->url().'/panel',
-      'text' => 'Panel'
+      'id'      => 'panel',
+      'icon'    => 'cogs',
+      'url'     => site()->url().'/panel',
+      'label'   => 'Panel',
+      'mobile'  => 'icon',
+    ));
+  }
+
+  protected function add() {
+    return self::link(array(
+      'id'     => 'add',
+      'icon'   => 'plus',
+      'url'    => $this->site->url().'/panel/#/pages/add/'.$this->page->uri(),
+      'label'  => 'Add',
+      'mobile' => 'icon',
     ));
   }
 
   protected function edit() {
     return self::link(array(
-      'id'   => 'edit',
-      'icon' => 'pencil-square-o',
-      'url'  => $this->site->url().'/panel/#/pages/show/'.$this->page->uri(),
-      'text' => 'Edit'
+      'id'     => 'edit',
+      'icon'   => 'pencil',
+      'url'    => $this->site->url().'/panel/#/pages/show/'.$this->page->uri(),
+      'label'  => 'Edit',
+      'mobile' => 'icon',
     ));
   }
 
   protected function toggle() {
     return self::link(array(
-      'id'   => 'toggle',
-      'icon' => $this->page->visible() ? 'toggle-on' : 'toggle-off',
-      'url'  => $this->site->url().'/panel/#/pages/toggle/'.$this->page->uri(),
-      'text' => $this->page->visible() ? 'Visible' : 'Invisible'
+      'id'     => 'toggle',
+      'icon'   => $this->page->isVisible() ? 'toggle-on' : 'toggle-off',
+      'url'    => $this->site->url().'/panel/#/pages/toggle/'.$this->page->uri(),
+      'label'  => $this->page->isVisible() ? 'Visible' : 'Invisible',
+      'mobile' => 'icon',
     ));
   }
 
   protected function user() {
     return self::link(array(
-      'id'    => 'user',
-      'icon'  => 'user',
-      'url'   => $this->site->url().'/panel/#/users/edit/'.$this->site->user(),
-      'text'  => $this->site->user(),
-      'float' => 'right'
+      'id'     => 'user',
+      'icon'   => 'user',
+      'url'    => $this->site->url().'/panel/#/users/edit/'.$this->site->user(),
+      'label'  => $this->site->user(),
+      'mobile' => 'icon',
+      'float'  => 'right',
     ));
   }
 
   protected function logout() {
     return self::link(array(
-      'id'    => 'logout',
-      'icon'  => 'power-off',
-      'url'   => $this->site->url().'/panel/logout',
-      'text'  => 'Logout',
-      'float' => 'right'
+      'id'     => 'logout',
+      'icon'   => 'power-off',
+      'url'    => $this->site->url().'/panel/logout',
+      'label'  => 'Logout',
+      'mobile' => 'icon',
+      'float'  => 'right',
     ));
   }
 
@@ -230,16 +284,17 @@ class PanelBar {
       $items = array();
       foreach($languages->not($this->site->language()->code()) as $language) {
         array_push($items, array(
-          'url' => $language->url().'/'.$this->page->uri(),
-          'text' => strtoupper($language->code())
+          'url'   => $language->url().'/'.$this->page->uri(),
+          'label' => strtoupper($language->code())
         ));
       }
 
       return self::dropdown(array(
-        'id'    => 'lang',
-        'icon'  => 'flag',
-        'label' => strtoupper($this->site->language()->code()),
-        'items' => $items
+        'id'     => 'lang',
+        'icon'   => 'flag',
+        'label'  => strtoupper($this->site->language()->code()),
+        'items'  => $items,
+        'mobile' => 'label'
       ));
     }
   }